# backend/openapi.json
{
  "components": {
    "schemas": {
      "BadgeResponse": {
        "properties": {
          "badgeKey": {
            "title": "Badgekey",
            "type": "string"
          },
          "earnedAt": {
            "title": "Earnedat",
            "type": "string"
          }
        },
        "required": [
          "badgeKey",
          "earnedAt"
        ],
        "title": "BadgeResponse",
        "type": "object"
      },
      "CompletionRequest": {
        "description": "day-level 완료 토글 — Android HomeScreen UX와 일치.\n\n하루를 통째로 완료 처리 (해당 date의 모든 exercises 일괄 완료 + day의 isCompleted 갱신).\n이전 exercise-level (day_index, exercise_index) 구조는 Android UX와 불일치해서 422 fail이었다.",
        "properties": {
          "completed": {
            "title": "Completed",
            "type": "boolean"
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "weekStart": {
            "title": "Weekstart",
            "type": "string"
          }
        },
        "required": [
          "weekStart",
          "date",
          "completed"
        ],
        "title": "CompletionRequest",
        "type": "object"
      },
      "GoalRequest": {
        "description": "PUT /goals — goal_type별 1개. 동일 type 재등록 시 target_value 갱신.",
        "properties": {
          "goalType": {
            "enum": [
              "weight",
              "body_fat"
            ],
            "title": "Goaltype",
            "type": "string"
          },
          "targetValue": {
            "exclusiveMinimum": 0.0,
            "maximum": 500.0,
            "title": "Targetvalue",
            "type": "number"
          }
        },
        "required": [
          "goalType",
          "targetValue"
        ],
        "title": "GoalRequest",
        "type": "object"
      },
      "GoalResponse": {
        "properties": {
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "goalType": {
            "title": "Goaltype",
            "type": "string"
          },
          "targetValue": {
            "title": "Targetvalue",
            "type": "number"
          }
        },
        "required": [
          "goalType",
          "targetValue",
          "createdAt"
        ],
        "title": "GoalResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ProfileHistoryEntry": {
        "properties": {
          "bodyFatPct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bodyfatpct"
          },
          "heightCm": {
            "title": "Heightcm",
            "type": "number"
          },
          "muscleMassKg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Musclemasskg"
          },
          "recordedAt": {
            "title": "Recordedat",
            "type": "string"
          },
          "weightKg": {
            "title": "Weightkg",
            "type": "number"
          }
        },
        "required": [
          "heightCm",
          "weightKg",
          "bodyFatPct",
          "muscleMassKg",
          "recordedAt"
        ],
        "title": "ProfileHistoryEntry",
        "type": "object"
      },
      "StatisticsResponse": {
        "description": "최근 N주(기본 12주) 완료율 + 현재/최장 스트릭.\n\n- weekly_rates: 주 단위 완료율, week_start 오름차순(차트용)\n- current_streak: 가장 최근 주부터 연속으로 100% 완료한 주 수\n- longest_streak: 전체 기간에서 100% 완료한 최장 연속 주 수",
        "properties": {
          "currentStreak": {
            "title": "Currentstreak",
            "type": "integer"
          },
          "longestStreak": {
            "title": "Longeststreak",
            "type": "integer"
          },
          "weeklyRates": {
            "items": {
              "$ref": "#/components/schemas/WeeklyRateDto"
            },
            "title": "Weeklyrates",
            "type": "array"
          }
        },
        "required": [
          "weeklyRates",
          "currentStreak",
          "longestStreak"
        ],
        "title": "StatisticsResponse",
        "type": "object"
      },
      "UserProfileRequest": {
        "properties": {
          "bodyFatPct": {
            "anyOf": [
              {
                "maximum": 70.0,
                "minimum": 1.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bodyfatpct"
          },
          "heightCm": {
            "maximum": 300.0,
            "minimum": 50.0,
            "title": "Heightcm",
            "type": "number"
          },
          "muscleMassKg": {
            "anyOf": [
              {
                "maximum": 200.0,
                "minimum": 1.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Musclemasskg"
          },
          "restDay": {
            "default": 7,
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Restday",
            "type": "integer"
          },
          "weightKg": {
            "maximum": 500.0,
            "minimum": 10.0,
            "title": "Weightkg",
            "type": "number"
          }
        },
        "required": [
          "heightCm",
          "weightKg"
        ],
        "title": "UserProfileRequest",
        "type": "object"
      },
      "UserProfileResponse": {
        "properties": {
          "bodyFatPct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bodyfatpct"
          },
          "heightCm": {
            "title": "Heightcm",
            "type": "number"
          },
          "muscleMassKg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Musclemasskg"
          },
          "restDay": {
            "default": 7,
            "title": "Restday",
            "type": "integer"
          },
          "userId": {
            "title": "Userid",
            "type": "string"
          },
          "weightKg": {
            "title": "Weightkg",
            "type": "number"
          }
        },
        "required": [
          "userId",
          "heightCm",
          "weightKg",
          "bodyFatPct",
          "muscleMassKg"
        ],
        "title": "UserProfileResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WeeklyPlanHistoryResponse": {
        "description": "페이지네이션 메타를 포함한 history envelope.\n\nAndroid의 HistoryViewModel이 total_count를 받아 페이지 인디케이터에 사용한다.\n이전엔 `list[WeeklyPlanResponse]`만 반환해서 Android Gson deserialization이 실패했다.",
        "properties": {
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "plans": {
            "items": {
              "$ref": "#/components/schemas/WeeklyPlanResponse"
            },
            "title": "Plans",
            "type": "array"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "totalCount": {
            "title": "Totalcount",
            "type": "integer"
          }
        },
        "required": [
          "plans",
          "totalCount",
          "page",
          "size"
        ],
        "title": "WeeklyPlanHistoryResponse",
        "type": "object"
      },
      "WeeklyPlanRequest": {
        "properties": {
          "dayPlans": {
            "title": "Dayplans",
            "type": "string"
          },
          "weekStart": {
            "title": "Weekstart",
            "type": "string"
          }
        },
        "required": [
          "weekStart",
          "dayPlans"
        ],
        "title": "WeeklyPlanRequest",
        "type": "object"
      },
      "WeeklyPlanResponse": {
        "properties": {
          "createdAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdat"
          },
          "dayPlans": {
            "title": "Dayplans",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "userId": {
            "title": "Userid",
            "type": "string"
          },
          "weekStart": {
            "title": "Weekstart",
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "weekStart",
          "dayPlans"
        ],
        "title": "WeeklyPlanResponse",
        "type": "object"
      },
      "WeeklyRateDto": {
        "description": "한 주의 운동 완료율 (0.0 ~ 1.0).",
        "properties": {
          "completionRate": {
            "title": "Completionrate",
            "type": "number"
          },
          "weekStart": {
            "title": "Weekstart",
            "type": "string"
          }
        },
        "required": [
          "weekStart",
          "completionRate"
        ],
        "title": "WeeklyRateDto",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "eundunHealth API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/.well-known/assetlinks.json": {
      "get": {
        "description": "Android App Links 검증용 assetlinks.json을 반환한다.",
        "operationId": "getAssetlinks",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Getassetlinks",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Assetlinks Json",
        "tags": [
          "auth"
        ]
      }
    },
    "/account": {
      "delete": {
        "description": "현재 인증 사용자의 계정을 삭제한다. Supabase Auth + 앱 DB 데이터 모두 제거.",
        "operationId": "deleteAccount",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Deleteaccount",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Account",
        "tags": [
          "account"
        ]
      }
    },
    "/auth/confirm": {
      "get": {
        "description": "앱 미설치 디바이스용 이메일 인증 완료 fallback 페이지를 반환한다.",
        "operationId": "getAuthConfirmFallback",
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Confirm Fallback",
        "tags": [
          "auth"
        ]
      }
    },
    "/badges": {
      "get": {
        "description": "현재 인증 사용자가 획득한 배지 목록을 반환한다.",
        "operationId": "getBadges",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BadgeResponse"
                  },
                  "title": "Response Getbadges",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Badges",
        "tags": [
          "badges"
        ]
      }
    },
    "/badges/{key}": {
      "post": {
        "description": "지정한 key 의 배지를 사용자에게 부여한다. 이미 보유하면 409 반환.",
        "operationId": "awardBadge",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadgeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Award Badge",
        "tags": [
          "badges"
        ]
      }
    },
    "/goals": {
      "get": {
        "description": "현재 인증 사용자의 목표 목록을 반환한다.",
        "operationId": "getGoals",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GoalResponse"
                  },
                  "title": "Response Getgoals",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Goals",
        "tags": [
          "goals"
        ]
      },
      "put": {
        "description": "목표를 생성하거나 갱신한다. 이미 존재하면 덮어쓴다 (upsert).",
        "operationId": "upsertGoal",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upsert Goal",
        "tags": [
          "goals"
        ]
      }
    },
    "/health": {
      "get": {
        "description": "서버 가동 상태를 확인한다. JWT 불필요 — probe 전용.",
        "operationId": "healthCheck",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Healthcheck",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "health"
        ]
      }
    },
    "/profile": {
      "get": {
        "description": "현재 인증 사용자의 프로필을 반환한다. 없으면 404.",
        "operationId": "getProfile",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Profile",
        "tags": [
          "profile"
        ]
      },
      "put": {
        "description": "프로필을 생성하거나 갱신한다. 변경 이력에 스냅샷을 자동 기록한다.",
        "operationId": "updateProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Updateprofile",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upsert Profile",
        "tags": [
          "profile"
        ]
      }
    },
    "/profile/history": {
      "get": {
        "description": "프로필 변경 이력 — 시간 오름차순 (차트 친화).",
        "operationId": "getProfileHistory",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProfileHistoryEntry"
                  },
                  "title": "Response Getprofilehistory",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Profile History",
        "tags": [
          "profile"
        ]
      }
    },
    "/weekly-plan": {
      "get": {
        "description": "지정한 주(week_start)의 운동 계획을 반환한다. 없으면 404.",
        "operationId": "getWeeklyPlan",
        "parameters": [
          {
            "in": "query",
            "name": "weekStart",
            "required": false,
            "schema": {
              "title": "Weekstart",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyPlanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Plan",
        "tags": [
          "weekly-plan"
        ]
      },
      "post": {
        "description": "주간 운동 계획을 생성하거나 교체한다. ExerciseDB 기반 자동 생성 결과를 저장한다.",
        "operationId": "createWeeklyPlan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeeklyPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyPlanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Plan",
        "tags": [
          "weekly-plan"
        ]
      }
    },
    "/weekly-plan/complete": {
      "patch": {
        "description": "특정 요일의 운동 완료 여부를 갱신한다. 모든 항목 완료 시 배지를 자동 부여한다.",
        "operationId": "updateDayCompletion",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Updatedaycompletion",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Complete",
        "tags": [
          "weekly-plan"
        ]
      }
    },
    "/weekly-plan/history": {
      "get": {
        "description": "과거 주간 계획 목록을 페이지 단위로 반환한다. 최신 순 정렬.",
        "operationId": "getWeeklyPlanHistory",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyPlanHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "History",
        "tags": [
          "weekly-plan"
        ]
      }
    },
    "/weekly-plan/previous": {
      "get": {
        "description": "기준 주 직전의 plan. 없으면 null — Android가 첫 사용자/공백 주를 구분하도록.",
        "operationId": "getPreviousWeeklyPlan",
        "parameters": [
          {
            "in": "query",
            "name": "weekStart",
            "required": false,
            "schema": {
              "title": "Weekstart",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/WeeklyPlanResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Getpreviousweeklyplan"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Previous",
        "tags": [
          "weekly-plan"
        ]
      }
    },
    "/weekly-plan/statistics": {
      "get": {
        "description": "최근 N주간 완료율과 연속 streak 통계를 반환한다.",
        "operationId": "getStatistics",
        "parameters": [
          {
            "in": "query",
            "name": "weeks",
            "required": false,
            "schema": {
              "default": 12,
              "maximum": 52,
              "minimum": 1,
              "title": "Weeks",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatisticsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Statistics",
        "tags": [
          "weekly-plan"
        ]
      }
    }
  }
}